refactor: split publish.ts into focused core/publish modules
Move adapters, Postiz, registry, and shared types out of the 1212-line
command file while re-exporting public symbols from commands/publish for
existing tests. Catalogs now aliases DEFAULT_BUILTIN_PUBLISH_ADAPTERS
from publish/builtins to avoid import cycles.
